ltc sync

+++ src/ltc/math/radix_to_bin.c
@@ -33,7 +33,7 @@
 */
 int radix_to_bin(const void *in, int radix, void *out, unsigned long *len)
 {
-   size_t l;
+   unsigned long l;
    void* mpi;
    int err;
